Try to avoid long unsigned int overflow. When free size is more than 4G, long unsigned int may enounter overflow
git-svn-id: http://opkg.googlecode.com/svn/trunk@210 e8e0d7a0-c8d9-11dd-a880-a1081c7ac358

+++ libopkg/opkg_utils.h
@@ -25,7 +25,7 @@ void push_error_list(struct errlist **errors,char * msg);
 void reverse_error_list(struct errlist **errors);
 void free_error_list();
 
-int get_available_blocks(char * filesystem);
+long unsigned int get_available_blocks(char * filesystem);
 char **read_raw_pkgs_from_file(const char *file_name);
 char **read_raw_pkgs_from_stream(FILE *fp);
 char *trim_alloc(char * line);
